MaterialView: split MaterialReservationCard into subcomponents
