+++ src/views/MaterialView/MaterialRulesDialog.js
