+++ src/views/MaterialView/index.js
@@ -4,16 +4,17 @@ import withStyles from '@material-ui/core/styles/withStyles';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 
-import viewStyle from '../assets/jss/viewStyle';
-import MaterialReservationCard from '../components/MaterialReservationCard';
-import MaterialReservationTable from '../components/MaterialReservationTable';
+import viewStyle from '../../assets/jss/viewStyle';
+import MaterialReservationForm from './MaterialReservationForm';
+import MaterialReservationTable from './MaterialReservationTable';
 
-import data from '../data.js';
+import data from '../../data.js';
+const reservations = data["material-reservations"];
+const material = data["material-items"];
+
 
 function MaterialView(props) {
   const { classes } = props;
-  const reservations = data["material-reservations"];
-  const names = data["material-items"];
 
   return (
       <main className={classes.mainContent}>
@@ -26,10 +27,18 @@ function MaterialView(props) {
         </Typography>
         <Grid container spacing={16}>
           <Grid item xs={12} sm={6} md={4}>
-            <MaterialReservationCard reservations={reservations} names={names}/>
+            <Typography variant="subtitle2">
+              Nieuwe reservatie
+            </Typography>
+            <MaterialReservationForm reservations={reservations} material={material}/>
           </Grid>
           <Grid item xs={12}>
-            <MaterialReservationTable reservations={reservations} />
+            <Typography variant="subtitle2">
+              Reservaties
+            </Typography>
+            <div style={{width: '100%', overflowX: 'auto'}}>
+              <MaterialReservationTable reservations={reservations} />
+            </div>
           </Grid>
         </Grid>
       </main>
